+++ packages/runtime/test/integration/group-inner-transaction.ts
@@ -2,32 +2,37 @@ import { types } from "@algo-builder/web";
 import { assert } from "chai";
 
 import { Runtime } from "../../src";
+import RUNTIME_ERRORS from "../../src/errors/errors-list";
 import { AccountStoreI, AppInfo } from "../../src/types";
 import { useFixture } from "../helpers/integration";
+import { expectRuntimeError } from "../helpers/runtime-errors";
 
 describe("Group inner transaction", function () {
 	useFixture("group-inner-transaction");
 
 	let runtime: Runtime;
-	let appInfo: AppInfo;
 	let alice: AccountStoreI;
 	let contractAcc: AccountStoreI;
 
-	function syncAccounts(): void {
+	let firstAppInfo: AppInfo;
+	let secondAppInfo: AppInfo;
+
+	function syncAccounts(appInfo: AppInfo): void {
 		contractAcc = runtime.getAccount(appInfo.applicationAccount);
 		[alice] = runtime.defaultAccounts();
 	}
-	this.beforeEach(() => {
-		runtime = new Runtime([]);
-		[alice] = runtime.defaultAccounts();
 
-		appInfo = runtime.deployApp(
+	function deployAppAndFund(
+		approvalProgramFilename: string,
+		clearProgramFilename: string
+	): AppInfo {
+		const appInfo = runtime.deployApp(
 			alice.account,
 			{
-				appName: "app",
+				appName: `${approvalProgramFilename}-${clearProgramFilename}`,
 				metaType: types.MetaType.FILE,
-				approvalProgramFilename: "group.py",
-				clearProgramFilename: "clear.teal",
+				approvalProgramFilename,
+				clearProgramFilename,
 				globalBytes: 1,
 				globalInts: 1,
 				localBytes: 1,
@@ -48,7 +53,16 @@ describe("Group inner transaction", function () {
 		};
 
 		runtime.executeTx([txnParams]);
-		syncAccounts();
+		syncAccounts(appInfo);
+
+		return appInfo;
+	}
+
+	this.beforeEach(() => {
+		runtime = new Runtime([]);
+		[alice] = runtime.defaultAccounts();
+
+		firstAppInfo = deployAppAndFund("group.py", "clear.teal");
 	});
 
 	it("Can use itxn_next for create group inner tx transaction", () => {
@@ -59,7 +73,7 @@ describe("Group inner transaction", function () {
 			type: types.TransactionType.CallApp,
 			sign: types.SignType.SecretKey,
 			fromAccount: alice.account,
-			appID: appInfo.appID,
+			appID: firstAppInfo.appID,
 			appArgs: ["str:call"],
 			payFlags: {
 				totalFee: 1000,
@@ -67,11 +81,52 @@ describe("Group inner transaction", function () {
 		};
 		assert.doesNotThrow(() => runtime.executeTx([txnParams]));
 
-		syncAccounts();
+		syncAccounts(firstAppInfo);
 
 		// receive 2000 micro Algo from contract and spend 1000 micro Algo to send transaction to network
 		assert.equal(aliceBalance + 1000n, alice.balance());
 		// spend 5000 micro Algo
 		assert.equal(contractBalance, contractAcc.balance() + 5000n);
+	});
+
+	describe("inner transaction limit tealv6", function () {
+		this.beforeEach(() => {
+			secondAppInfo = deployAppAndFund("limit-number-txn.py", "clear.teal");
+		});
+
+		it("Should fail when issue more than 256 inner txn", () => {
+			const firstTxn: types.ExecParams = {
+				type: types.TransactionType.CallApp,
+				sign: types.SignType.SecretKey,
+				fromAccount: alice.account,
+				appID: secondAppInfo.appID,
+				appArgs: ["str:exec"],
+				payFlags: {},
+			};
+
+			const secondTxn: types.ExecParams = {
+				...firstTxn,
+				payFlags: {
+					note: "second",
+				},
+			};
+			const thirdTxn: types.ExecParams = {
+				...firstTxn,
+				payFlags: {
+					note: "third",
+				},
+			};
+			const fourthTxn: types.ExecParams = {
+				...firstTxn,
+				payFlags: {
+					note: "fourth",
+				},
+			};
+
+			expectRuntimeError(
+				() => runtime.executeTx([firstTxn, secondTxn, thirdTxn, fourthTxn]),
+				RUNTIME_ERRORS.GENERAL.TOO_MANY_INNER_TXN
+			);
+		});
 	});
 });